Refacotr: Rename components
